+++ src/main/java/com/teamtea/eclipticseasons/api/constant/tag/EclipticBlockTags.java
@@ -18,6 +18,7 @@ public class EclipticBlockTags {
     public static final TagKey<Block> DARK_GROW_PLANTS = create("dark_grow_plants");
 
     public static final TagKey<Block> NATURAL_PLANTS = create("natural_plants");
+    public static final TagKey<Block> VOLATILE = create("volatile");
     public static final TagKey<Block> VOLATILE_PLANTS = create("volatile_plants");
 
     public static TagKey<Block> create(String s) {