+++ src/main/java/com/teamtea/eclipticseasons/api/data/misc/PosAndBlockStateCheck.java
@@ -2,6 +2,7 @@ package com.teamtea.eclipticseasons.api.data.misc;
 
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
+import com.teamtea.eclipticseasons.api.util.backport.FakeBlockPredicate;
 import com.teamtea.eclipticseasons.api.util.codec.ESExtraCodec;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.HolderSet;
@@ -11,16 +12,16 @@ import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
 
 public record PosAndBlockStateCheck(
-        Vec3i offset, HolderSet<Block> block
+        Vec3i offset, FakeBlockPredicate block
 ) {
 
     public static final Codec<PosAndBlockStateCheck> CODEC = RecordCodecBuilder.create(builder -> builder.group(
             Vec3i.CODEC.fieldOf("offset").forGetter(PosAndBlockStateCheck::offset),
-            ESExtraCodec.BLOCK_HOLDER_SET_CODEC.fieldOf("block").forGetter(PosAndBlockStateCheck::block)
+            FakeBlockPredicate.CODEC.fieldOf("block").forGetter(PosAndBlockStateCheck::block)
     ).apply(builder, PosAndBlockStateCheck::new));
 
     private boolean matchesState(BlockState state) {
-        return block().contains(state.getBlockHolder());
+        return block().matches(state);
     }
 
     public boolean matches(ServerLevel level, BlockPos pos) {