+++ src/main/java/com/teamtea/eclipticseasons/api/misc/CustomRandomTick.java
@@ -5,7 +5,8 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.block.state.BlockState;
 
-public interface CustomRandomTick
-{
-    void tick(BlockState state, ServerLevel worldIn, BlockPos pos);
+public interface CustomRandomTick {
+    void eclipticseasons$tick(BlockState state, ServerLevel worldIn, BlockPos pos);
+
+    void eclipticseasons$reset();
 }