+++ src/main/java/com/teamtea/eclipticseasons/api/util/backport/FakeBlockPredicate.java
@@ -2,10 +2,15 @@ package com.teamtea.eclipticseasons.api.util.backport;
 
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
+import net.minecraft.core.BlockPos;
 import net.minecraft.core.HolderSet;
 import net.minecraft.core.RegistryCodecs;
 import net.minecraft.core.registries.Registries;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.state.BlockState;
 
 import java.util.Optional;
 
@@ -17,4 +22,22 @@ public record FakeBlockPredicate(Optional<HolderSet<Block>> blocks, Optional<Fak
                     )
                     .apply(p_337342_, FakeBlockPredicate::new)
     );
+
+    public FakeBlockPredicate(HolderSet<Block> blockHolderSet) {
+        this(Optional.of(blockHolderSet), Optional.empty());
+    }
+
+    public boolean matches(BlockState state) {
+        return (this.blocks.isEmpty() || state.is(this.blocks.get()))
+                && (this.properties.isEmpty() || this.properties.get().matches(state));
+    }
+
+    public boolean matches(Level level, BlockPos blockPos) {
+        return matches(level.getBlockState(blockPos));
+    }
+
+
+    public HolderSet<Block> getBlocks() {
+        return blocks.isPresent() ? blocks.get() : HolderSet.direct();
+    }
 }