+++ src/main/java/com/teamtea/eclipticseasons/data/general/tag/ESBlockTagProvider.java
@@ -34,8 +34,8 @@ public final class ESBlockTagProvider extends BlockTagsProvider {
                 .addOptional(fd_rl("red_mushroom_colony"));
 
         tag(EclipticBlockTags.NATURAL_PLANTS).add(Blocks.BAMBOO_SAPLING);
+        tag(EclipticBlockTags.VOLATILE).add(Blocks.BUBBLE_COLUMN).addTag(EclipticBlockTags.VOLATILE_PLANTS);
         tag(EclipticBlockTags.VOLATILE_PLANTS);
-
         // add crop info
         tag(CropSeasonType.SUMMER.getBlockTag()).add(Blocks.MELON_STEM, Blocks.COCOA, Blocks.CACTUS);
         tag(CropSeasonType.AUTUMN.getBlockTag()).add(Blocks.PUMPKIN_STEM);